Moved all exceptions to the single file in core folder
