+++ core/exceptions.py
