+++ core/services/pvp.py
@@ -2,7 +2,7 @@ from datetime import datetime, timedelta
 from typing import Final
 
 from core.base_bot import BaseBotAPI
-from core.exceptions.pvb import (
+from core.exceptions import (
     BetOutOfLimitsError,
     BalanceIsNotEnoughError,
 )
